requirements.txt
Django==1.7.1
Jinja2==2.7.3
Pillow==2.6.1
django-jinja==1.0.4
PyYAML==3.11
dj-static==0.0.6
gunicorn==19.0.0
psycopg2==2.5.4
easy-thumbnails==2.2
django-haystack==2.3.1
Whoosh==2.6.0